--- include/phase4/common/game_phase.h
@@ -1,4 +1,0 @@
-#ifndef PHASE4_COMMON_GAME_PHASE_H
-#define PHASE4_COMMON_GAME_PHASE_H
-
-#endif
