--- include/phase4/common/position.h
@@ -1,4 +1,0 @@
-#ifndef PHASE4_COMMON_POSITION_H
-#define PHASE4_COMMON_POSITION_H
-
-#endif
