--- include/phase4/common/wall_operations.h
@@ -1,4 +1,0 @@
-#ifndef PHASE4_COMMON_WALL_OPERATIONS_H
-#define PHASE4_COMMON_WALL_OPERATIONS_H
-
-#endif
