+++ include/phase4/engine/common/direction.h
@@ -1,0 +1,4 @@
+#ifndef PHASE4_ENGINE_COMMON_DIRECTION_H
+#define PHASE4_ENGINE_COMMON_DIRECTION_H
+
+#endif
