+++ include/phase4/engine/common/game_phase.h
@@ -1,0 +1,4 @@
+#ifndef PHASE4_ENGINE_COMMON_GAME_PHASE_H
+#define PHASE4_ENGINE_COMMON_GAME_PHASE_H
+
+#endif
