+++ include/phase4/engine/common/position.h
@@ -1,0 +1,4 @@
+#ifndef PHASE4_ENGINE_COMMON_POSITION_H
+#define PHASE4_ENGINE_COMMON_POSITION_H
+
+#endif
