+++ include/phase4/engine/common/wall_operations.h
@@ -1,0 +1,4 @@
+#ifndef PHASE4_ENGINE_COMMON_WALL_OPERATIONS_H
+#define PHASE4_ENGINE_COMMON_WALL_OPERATIONS_H
+
+#endif
