+++ include/phase4/engine/moves/move.h
@@ -1,0 +1,4 @@
+#ifndef PHASE4_ENGINE_MOVES_MOVE_H
+#define PHASE4_ENGINE_MOVES_MOVE_H
+
+#endif
