--- include/phase4/moves/move.h
@@ -1,4 +1,0 @@
-#ifndef PHASE4_MOVES_MOVE_H
-#define PHASE4_MOVES_MOVE_H
-
-#endif
