--- include/phase4/moves/move_flags.h
@@ -1,4 +1,0 @@
-#ifndef PHASE4_MOVES_MOVE_FLAGS_H
-#define PHASE4_MOVES_MOVE_FLAGS_H
-
-#endif
