+++ src/uci/uci.m.cpp
@@ -1,0 +1,2 @@
+int main(int argc, char *argv[]) {
+}
