+++ tests/phase4/engine/common/bitset.t.cpp
@@ -1,4 +1,4 @@
-#include <phase4/common/bitset.h>
+#include <phase4/engine/common/bitset.h>
 
 #include <sstream>
